Configured admin-page to list all products from database
