+++ app/components/RenderProducts.tsx
@@ -1,0 +1,10 @@
+export default async function RenderProducts() {
+  //   try {
+  //     const products = await getAllProducts();
+  //     console.log(products);
+  //     return;
+  //   } catch (error) {
+  //     console.log(error);
+  //   }
+  return <div>RenderProducts</div>;
+}
