+++ app/types.ts
@@ -2,3 +2,8 @@ import { ReactNode } from "react";
 
 export type PageProps = Readonly<{ params: { slug: string } }>;
 export type LayoutProps = Readonly<{ children: ReactNode }>;
+
+// export interface Product {
+//   createdAt: Date;
+//   desc: String;
+// }
